server/requirements.txt
certifi==2019.3.9
chardet==3.0.4
grpcio==1.21.1
grpcio-tools==1.21.1
idna==2.8
protobuf==3.8.0
PyCap==1.0.0
requests==2.22.0
semantic-version==2.3.1
six==1.12.0
urllib3==1.25.3
